feat(seo): fix discoverability gaps with page-specific titles, preview text, and images
Replace generic fallback descriptions and URL-based titles across wiki articles and bespoke routes with dynamic frontmatter/H1 titles, first-paragraph summaries, and resolved OpenGraph/Twitter card preview images.
