+++ src/app/[locale]/developers/page.tsx
@@ -1,28 +1,32 @@
-import { Metadata } from 'next';
-import { genMetadata } from '@/lib/helpers';
-import { buildAlternates } from '@/lib/localeCoverage';
-import { routing } from '@/i18n/routing';
-import DeveloperPage from './DeveloperPage';
-import { getDictionary } from '@/lib/getDictionary';
+import { Metadata } from "next";
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { buildAlternates } from "@/lib/localeCoverage";
+import { routing } from "@/i18n/routing";
+import DeveloperPage from "./DeveloperPage";
+import { getDictionary } from "@/lib/getDictionary";
 
 type DevelopersDictionary = {
   pages?: {
     developers?: {
       title?: string;
+      description?: string;
     };
   };
 };
 
 export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }): Promise<Metadata> {
   const { locale } = await params;
-  const dict = (await getDictionary(locale)) as DevelopersDictionary;
+  const dict = (await getDictionary(locale).catch(() => ({}))) as DevelopersDictionary;
   const localePrefix = locale && locale !== routing.defaultLocale ? `/${locale}` : "";
-  // Bespoke app route: English-only in the sitemap, so emit a locale-aware self
-  // canonical (no hreflang). (Also fixes a prior copy-paste bug: the url pointed
-  // at /using-zcash/blockchain-explorers.)
+  const dev = dict.pages?.developers;
+
   return genMetadata({
-    title: dict.pages?.developers?.title || "Zcash Developer Resources",
+    title: dev?.title ? `${dev.title} | ZecHub` : "Zcash Developer Hub, SDKs & APIs | ZecHub",
+    description:
+      dev?.description ??
+      "Comprehensive developer documentation, SDKs, light client libraries, node software, and API resources for building on Zcash.",
     url: `https://zechub.wiki${localePrefix}/developers`,
+    image: getBanner("developers") || "/content-banners/bannertech.jpg",
     locale,
     alternates: buildAlternates("/developers", locale, [locale]),
   });
