+++ src/app/[locale]/developers/quick-start/QuickStartClient.tsx
@@ -1,0 +1,1271 @@
+"use client";
+
+import React, { useState, useRef, useEffect, ReactNode } from "react";
+import { FcLinux } from "react-icons/fc";
+import { PiAppleLogoDuotone } from "react-icons/pi";
+import { BsWindows } from "react-icons/bs";
+import { useLanguage } from "@/context/LanguageContext";
+
+// ── Types ────────────────────────────────────────────────────────────────────
+
+interface CodeBlockProps {
+  code: string;
+  language?: string;
+  width?: string;
+}
+
+interface SectionCardProps {
+  title: string;
+  children: ReactNode;
+  className?: string;
+}
+
+interface StepBadgeProps {
+  number: number;
+}
+
+// ── Utilities ────────────────────────────────────────────────────────────────
+
+const CopyIcon = () => (
+  <svg
+    width="14"
+    height="14"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
+    <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
+  </svg>
+);
+
+const CheckIcon = () => (
+  <svg
+    width="14"
+    height="14"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2.5"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <polyline points="20 6 9 17 4 12" />
+  </svg>
+);
+
+const ExternalLinkIcon = () => (
+  <svg
+    width="12"
+    height="12"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" />
+    <polyline points="15 3 21 3 21 9" />
+    <line x1="10" y1="14" x2="21" y2="3" />
+  </svg>
+);
+
+// ── Components ───────────────────────────────────────────────────────────────
+
+// CodeBlock intentionally stays dark (terminal aesthetic) in both modes
+const CodeBlock = ({
+  code,
+  language = "bash",
+  width = "850px",
+}: CodeBlockProps) => {
+  const { t } = useLanguage();
+  const [copied, setCopied] = useState(false);
+
+  const handleCopy = async () => {
+    await navigator.clipboard.writeText(code.trim());
+    setCopied(true);
+    setTimeout(() => setCopied(false), 2000);
+  };
+
+  return (
+    <div
+      style={{
+        position: "relative",
+        background: "#0a0e17",
+        border: "1px solid #1e293b",
+        borderRadius: "8px",
+        overflow: "hidden",
+        fontFamily: "'JetBrains Mono', 'Fira Code', monospace",
+      }}
+    >
+      {/* Terminal header bar */}
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "space-between",
+          padding: "8px 14px",
+          background: "#0d1117",
+          borderBottom: "1px solid #1e293b",
+        }}
+      >
+        <div style={{ display: "flex", gap: "6px" }}>
+          <span
+            style={{
+              width: 10,
+              height: 10,
+              borderRadius: "50%",
+              background: "#ff5f57",
+              display: "block",
+            }}
+          />
+          <span
+            style={{
+              width: 10,
+              height: 10,
+              borderRadius: "50%",
+              background: "#febc2e",
+              display: "block",
+            }}
+          />
+          <span
+            style={{
+              width: 10,
+              height: 10,
+              borderRadius: "50%",
+              background: "#28c840",
+              display: "block",
+            }}
+          />
+        </div>
+        <span
+          style={{
+            fontSize: "10px",
+            color: "#4a5568",
+            letterSpacing: "0.1em",
+            textTransform: "uppercase",
+          }}
+        >
+          {language}
+        </span>
+        <button
+          onClick={handleCopy}
+          style={{
+            display: "flex",
+            alignItems: "center",
+            gap: "5px",
+            padding: "4px 10px",
+            background: copied
+              ? "rgba(34, 197, 94, 0.15)"
+              : "rgba(255,255,255,0.06)",
+            border: `1px solid ${copied ? "rgba(34,197,94,0.4)" : "rgba(255,255,255,0.1)"}`,
+            borderRadius: "5px",
+            color: copied ? "#22c55e" : "#9ca3af",
+            fontSize: "11px",
+            fontFamily: "inherit",
+            cursor: "pointer",
+            transition: "all 0.2s ease",
+            letterSpacing: "0.02em",
+          }}
+        >
+          {copied ? <CheckIcon /> : <CopyIcon />}
+          {copied
+            ? (t?.pages?.developersQuickStart?.copied ?? "Copied!")
+            : (t?.pages?.developersQuickStart?.copy ?? "Copy")}
+        </button>
+      </div>
+
+      {/* Code content */}
+      <pre
+        style={{
+          margin: 0,
+          padding: "18px 20px",
+          overflowX: "auto",
+          fontSize: "13px",
+          lineHeight: "1.75",
+          color: "#e2e8f0",
+        }}
+      >
+        <code style={{ maxWidth: width }}>{code.trim()}</code>
+      </pre>
+    </div>
+  );
+};
+
+const SectionCard = ({ title, children, className = "" }: SectionCardProps) => (
+  <div
+    className={`${className} p-4 rounded-xl border border-black/8 dark:border-white/8 bg-black/[0.02] dark:bg-white/[0.03] backdrop-blur-sm`}
+  >
+    <h4
+      style={{
+        fontSize: "13px",
+        fontWeight: 600,
+        letterSpacing: "0.12em",
+        textTransform: "uppercase",
+        color: "#22d3ee",
+        marginBottom: "18px",
+        display: "flex",
+        alignItems: "center",
+        gap: "8px",
+      }}
+    >
+      <span
+        style={{
+          width: 6,
+          height: 6,
+          borderRadius: "50%",
+          background: "#22d3ee",
+          display: "inline-block",
+        }}
+      />
+      {title}
+    </h4>
+    {children}
+  </div>
+);
+
+const StepBadge = ({ number }: StepBadgeProps) => (
+  <span
+    style={{
+      display: "inline-flex",
+      alignItems: "center",
+      justifyContent: "center",
+      width: 32,
+      height: 32,
+      borderRadius: "50%",
+      background: "linear-gradient(135deg, #0ea5e9, #6366f1)",
+      color: "#fff",
+      fontSize: "13px",
+      fontWeight: 700,
+      flexShrink: 0,
+      boxShadow: "0 0 20px rgba(14,165,233,0.3)",
+    }}
+  >
+    {number}
+  </span>
+);
+
+const Divider = () => (
+  <div className="my-9 h-px bg-gradient-to-r from-transparent via-black/10 dark:via-white/8 to-transparent" />
+);
+
+const ResourceLink = ({
+  href,
+  children,
+}: {
+  href: string;
+  children: ReactNode;
+}) => (
+  <a
+    href={href}
+    target="_blank"
+    rel="noopener noreferrer"
+    className="flex items-center gap-2 px-[14px] py-[10px] rounded-lg text-[13.5px] no-underline transition-all duration-200
+      border border-black/8 dark:border-white/8
+      bg-black/[0.02] dark:bg-white/[0.04]
+      text-slate-500 dark:text-slate-400
+      hover:bg-sky-500/8 hover:border-sky-500/30 hover:text-slate-700 dark:hover:text-slate-200"
+  >
+    <ExternalLinkIcon />
+    {children}
+  </a>
+);
+
+// ── Shared sub-labels ────────────────────────────────────────────────────────
+
+const SubLabel = ({
+  children,
+  className = "",
+}: {
+  children: ReactNode;
+  className?: string;
+}) => (
+  <p
+    className={`text-[11px] tracking-[0.1em] uppercase text-slate-900 dark:text-slate-600 mb-2 ${className}`}
+  >
+    {children}
+  </p>
+);
+
+const InfoBox = ({ children }: { children: ReactNode }) => (
+  <div
+    className="rounded-[10px] p-[18px_22px] mb-6 text-sm leading-relaxed
+      bg-indigo-500/[0.06] dark:bg-indigo-500/[0.08]
+      border border-indigo-500/20
+      text-slate-600 dark:text-slate-400"
+  >
+    {children}
+  </div>
+);
+
+// ── Tab content ──────────────────────────────────────────────────────────────
+
+const ZebradTab = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+  const [osTab, setOsTab] = useState<"linux" | "macos" | "windows">("linux");
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "40px" }}>
+      {/* TOC */}
+      <div className="rounded-xl px-7 py-6 bg-sky-500/[0.04] dark:bg-sky-500/[0.06] border border-sky-500/20">
+        <p className="text-[11px] tracking-[0.15em] uppercase text-sky-500 mb-[14px] font-semibold">
+          {qs?.onThisPage ?? "On this page"}
+        </p>
+        <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
+          {[
+            [
+              "#installing-zcash",
+              qs?.tocInstallingZebrad ?? "Installing Zebrad",
+            ],
+            ["#running-zebrad", qs?.tocRunningZebrad ?? "Running zebrad"],
+            [
+              "#connecting-lightwalletd",
+              qs?.tocConnectingLightwalletd ?? "Connecting with lightwalletd",
+            ],
+          ].map(([href, label]) => (
+            <a
+              key={href}
+              href={href}
+              className="text-slate-900 dark:text-slate-400 hover:text-slate-600 dark:hover:text-slate-200 text-sm no-underline flex items-center gap-2 max-w-fit transition-colors"
+            >
+              <span className="w-1 h-1 rounded-full bg-slate-400 dark:bg-slate-600 inline-block" />
+              {label}
+            </a>
+          ))}
+        </div>
+      </div>
+
+      {/* Step 1 — Installing */}
+      <SectionCard title={qs?.installation ?? "Installation"}>
+        {/* OS Tab bar */}
+        <div className="grid grid-cols-2 imd:flex gap-1 border-b border-black/8 dark:border-white/8 mb-6">
+          {(
+            [
+              { id: "linux", icon: FcLinux, label: "Linux", color: null },
+              {
+                id: "macos",
+                icon: PiAppleLogoDuotone,
+                label: "macOS",
+                color: null,
+              },
+              {
+                id: "windows",
+                icon: BsWindows,
+                label: "Windows",
+                color: "#22d3ee",
+              },
+            ] as const
+          ).map(({ id, icon: Icon, label, color }) => (
+            <button
+              key={id}
+              onClick={() => setOsTab(id)}
+              className={`flex items-center gap-2 px-[18px] py-[10px] bg-transparent border-0 border-b-2 cursor-pointer transition-all duration-200 text-sm ${
+                osTab === id
+                  ? "border-b-sky-500 text-slate-900 dark:text-slate-200 font-semibold"
+                  : "border-b-transparent text-slate-400 dark:text-slate-500 font-normal hover:text-slate-600 dark:hover:text-slate-300"
+              }`}
+            >
+              <Icon size={18} color={color ?? ""} />
+              {label}
+            </button>
+          ))}
+        </div>
+
+        {osTab === "linux" && (
+          <>
+            <SubLabel className="mt-[14px]">
+              {qs?.dependencies ?? "Dependencies"}
+            </SubLabel>
+            <CodeBlock
+              code={`curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh\nsudo apt update\nsudo apt install libclang-dev clang pkg-config openssl protobuf-compiler npm`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.fromSource ?? "From Source"}
+            </SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/ZcashFoundation/zebra.git\ncd zebra\ngit checkout v6.3.0\ncargo build --release --bin zebrad\nexport PATH="$PATH:(pwd)/target/release"`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.alternatively ?? "Alternatively"}
+            </SubLabel>
+            <CodeBlock
+              code={`cargo install --git https://github.com/ZcashFoundation/zebra --tag v6.3.0 zebrad`}
+            />
+          </>
+        )}
+
+        {osTab === "macos" && (
+          <>
+            <SubLabel className="mt-[14px]">
+              {qs?.dependencies ?? "Dependencies"}
+            </SubLabel>
+            <CodeBlock
+              code={`curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh\nsudo apt update\nsudo apt install libclang-dev clang pkg-config openssl protobuf-compiler npm`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.manualDownload ?? "Manual Download"}
+            </SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/ZcashFoundation/zebra.git\ncd zebra\ngit checkout v6.3.0\ncargo build --release --bin zebrad\nexport PATH="$PATH:(pwd)/target/release"`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.alternatively ?? "Alternatively"}
+            </SubLabel>
+            <CodeBlock
+              code={`cargo install --git https://github.com/ZcashFoundation/zebra --tag v6.3.0 zebrad`}
+            />
+          </>
+        )}
+
+        {osTab === "windows" && (
+          <>
+            <SubLabel>{qs?.wslRecommended ?? "WSL (Recommended)"}</SubLabel>
+            <CodeBlock code={`wsl --install`} />
+            <SubLabel className="mt-[14px]">
+              {qs?.dependencies ?? "Dependencies"}
+            </SubLabel>
+            <CodeBlock
+              code={`curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh\nsudo apt update\nsudo apt install libclang-dev clang pkg-config openssl protobuf-compiler npm\nsource $HOME/.cargo/env\nexport CXXFLAGS="$CXXFLAGS -include cstdint"`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.fromSource ?? "From Source"}
+            </SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/ZcashFoundation/zebra.git\ncd zebra\ngit checkout v6.3.0\ncargo build --release --bin zebrad\nexport PATH="$PATH:(pwd)/target/release"`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.alternatively ?? "Alternatively"}
+            </SubLabel>
+            <CodeBlock
+              code={`cargo install --git https://github.com/ZcashFoundation/zebra --tag v6.3.0 zebrad`}
+            />
+          </>
+        )}
+      </SectionCard>
+
+      <Divider />
+
+      {/* Step 2 — Running */}
+      <section id="running-zebrad">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={2} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.runningZebradHeading ?? "Running zebrad"}
+          </h2>
+        </div>
+
+        <SectionCard
+          title={qs?.initialSetup ?? "Initial Setup"}
+          className="mb-6"
+        >
+          <SubLabel>
+            {qs?.createConfigDir ?? "Create Configuration Directory"}
+          </SubLabel>
+          <CodeBlock code={`zebrad generate -o ~/.config/zebrad.toml`} />
+          <SubLabel className="mt-[18px]">zebrad.toml</SubLabel>
+          <CodeBlock
+            language="toml"
+            code={`[consensus]
+checkpoint_sync = true
+
+[mempool]
+eviction_memory_time = "1h"
+tx_cost_limit = 80000000
+
+[network]
+cache_dir = true
+crawl_new_peer_interval = "1m 1s"
+listen_addr = "[::]:8233"
+network = "Mainnet"
+peerset_initial_target_size = 25
+
+[rpc]
+listen_addr = "127.0.0.1:8232"
+cookie_dir = "/home/your_username/.cache/zebra"
+enable_cookie_auth = false
+
+[state]
+cache_dir = "/home/your_username/.cache/zebra" # (Change to store in external SSD)
+delete_old_database = true
+ephemeral = false
+
+[sync]
+checkpoint_verify_concurrency_limit = 1000
+download_concurrency_limit = 50
+full_verify_concurrency_limit = 20
+
+[tracing]
+buffer_limit = 128000
+use_color = true`}
+          />
+        </SectionCard>
+
+        <div
+          className="grid gap-5"
+          style={{
+            gridTemplateColumns: "repeat(auto-fit, minmax(280px, 1fr))",
+          }}
+        >
+          <SectionCard title={qs?.startingZebrad ?? "Starting zebrad"}>
+            <SubLabel>{qs?.commandLine ?? "Command Line"}</SubLabel>
+            <CodeBlock
+              code={`zebrad start\n# Or with specific config\nzebrad -c ~/.config/zebrad.toml start`}
+            />
+          </SectionCard>
+
+          <SectionCard title={qs?.importantNotes ?? "Important Notes"}>
+            <ul className="flex flex-col gap-[10px] list-none p-0 m-0">
+              {[
+                qs?.noteDownloadsBlockchain ??
+                  "First run downloads the entire blockchain (~250 GB)",
+                qs?.noteInitialSync ?? "Initial sync can take several hours",
+                qs?.noteKeepSecure ??
+                  "Keep your zebrad.toml secure and private",
+              ].map((note, i) => (
+                <li
+                  key={i}
+                  className="flex gap-[10px] text-[13.5px] text-slate-500 dark:text-slate-400 leading-relaxed"
+                >
+                  <span className="text-amber-500 shrink-0 mt-[2px]">⚠</span>
+                  {note}
+                </li>
+              ))}
+            </ul>
+          </SectionCard>
+        </div>
+      </section>
+
+      <Divider />
+
+      {/* Step 3 — lightwalletd */}
+      <section id="connecting-lightwalletd">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={3} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.connectingLightwalletdHeading ??
+              "Connecting with lightwalletd"}
+          </h2>
+        </div>
+
+        <InfoBox>
+          <strong className="text-indigo-400 dark:text-indigo-300">
+            lightwalletd
+          </strong>{" "}
+          {qs?.lightwalletdInfo ??
+            "is a server that provides a lightweight interface to the Zcash blockchain, designed for mobile and desktop wallets that don't need to run a full node."}
+        </InfoBox>
+
+        <div
+          className="grid grid-cols-1 imd:grid-cols-2 gap-5"
+        >
+          <SectionCard title={qs?.installation ?? "Installation"}>
+            <SubLabel>{qs?.prerequisites ?? "Prerequisites"}</SubLabel>
+            <CodeBlock
+              code={`# Install Go
+sudo apt install golang-go
+# Or download from golang.org`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.buildLightwalletd ?? "Build lightwalletd"}
+            </SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/zcash/lightwalletd.git\ncd lightwalletd\ngo build\nmake\nmake install\nexport PATH=$PATH:~/go/bin`}
+            />
+          </SectionCard>
+
+          <SectionCard title={qs?.configuration ?? "Configuration"}>
+            <SubLabel>
+              {qs?.basicSetupMainnet ?? "Basic Setup (Mainnet)"}
+            </SubLabel>
+            <CodeBlock
+              code={`cat > ~/.config/zcash.conf << EOF
+rpcport=8232
+rpcbind=127.0.0.1
+rpcuser=your_rpc_username
+rpcpassword=your_rpc_password
+EOF`}
+            />
+            <SubLabel className="pt-4">
+              {qs?.basicSetupTestnet ?? "Basic Setup (Testnet)"}
+            </SubLabel>
+            <CodeBlock
+              code={`cat > ~/.config/zcash.conf << EOF
+testnet=1
+rpcport=18238
+rpcbind=127.0.0.1
+rpcuser=your_rpc_username
+rpcpassword=your_rpc_password
+EOF`}
+            />
+          </SectionCard>
+        </div>
+
+        <SectionCard className="mt-6" title={qs?.run ?? "Run"}>
+          <CodeBlock
+            width="fit-content"
+            code={`lightwalletd --zcash-conf-path ~/.config/zcash.conf --data-dir ~/data/zebrad/.cache/lightwalletd --log-file ~/.local/state/lwd.log --no-tls-very-insecure`}
+          />
+        </SectionCard>
+      </section>
+
+      <ResourcesSection />
+    </div>
+  );
+};
+
+const ZainodTab = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "40px" }}>
+      <section id="installing-zaino">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={1} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.installingZaino ?? "Installing Zaino"}
+          </h2>
+        </div>
+
+        <InfoBox>
+          <strong className="text-indigo-400 dark:text-indigo-300">
+            Zaino
+          </strong>{" "}
+          {qs?.zainoInfo ??
+            "is a Rust-based indexer for the Zcash blockchain. It serves both light clients (wallets that don't store the full history) and full clients / block explorers, providing access to finalized chain, the non-finalized best chain, and the mempool."}
+        </InfoBox>
+
+        <div
+          className="grid gap-5"
+          style={{
+            gridTemplateColumns: "repeat(auto-fit, minmax(280px, 1fr))",
+          }}
+        >
+          <SectionCard title={qs?.installation ?? "Installation"}>
+            <SubLabel>{qs?.buildZaino ?? "Build Zaino"}</SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/zingolabs/zaino.git\ncd zaino\ncargo build --release\nPATH=$PATH:~/zaino/target/release/`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.configure ?? "Configure"}
+            </SubLabel>
+            <CodeBlock
+              code={`cd ~/zaino/zainod\nsudo nano zindexer.toml\n# Adjust port to 8232 for mainnet`}
+            />
+          </SectionCard>
+
+          <SectionCard title={qs?.configuration ?? "Configuration"}>
+            <SubLabel>zindexer.toml</SubLabel>
+            <CodeBlock
+              language="toml"
+              code={`# TcpIngestor status
+tcp_active = true
+
+# Listen port
+listen_port = 8137
+
+# LightWalletD listen port [DEPRECATED]
+lightwalletd_port = 9067
+
+# Full node / validator listen port
+zebrad_port = 8232
+
+# Optional credentials
+node_user = "xxxxxx"
+node_password = "xxxxxx"
+
+# Worker pool settings
+max_queue_size = 1024
+max_worker_pool_size = 64
+idle_worker_pool_size = 4`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.runZainod ?? "Run zainod"}
+            </SubLabel>
+            <CodeBlock code={`zainod --config zindexer.toml`} />
+          </SectionCard>
+        </div>
+      </section>
+
+      <ResourcesSection />
+    </div>
+  );
+};
+
+const ZingolibTab = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "40px" }}>
+      <section id="installing-zingo">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={1} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.installingZingoCli ?? "Installing Zingo CLI"}
+          </h2>
+        </div>
+
+        <InfoBox>
+          <strong className="text-indigo-400 dark:text-indigo-300">
+            Zingo-cli
+          </strong>{" "}
+          {qs?.zingoCliInfo ??
+            "is a command-line lightwalletd-proxy client built in Rust. Releases are currently provisional — this guide covers compiling from source."}
+        </InfoBox>
+
+        <div
+          className="grid gap-5"
+          style={{
+            gridTemplateColumns: "repeat(auto-fit, minmax(280px, 1fr))",
+          }}
+        >
+          <SectionCard title={qs?.installation ?? "Installation"}>
+            <SubLabel>{qs?.buildZingoCli ?? "Build Zingo CLI"}</SubLabel>
+            <CodeBlock
+              code={`git clone https://github.com/zingolabs/zingolib.git\ncd zingolib\ngit checkout tags/zingolib_v5.0.0\ncargo build --release --package zingo-cli\ncp target/release/zingo-cli ~/.cargo/bin/`}
+            />
+          </SectionCard>
+
+          <SectionCard title={qs?.running ?? "Running"}>
+            <div className="pb-4">
+              <SubLabel>
+                {qs?.startZingoCliDefault ?? "Start Zingo CLI (Default)"}
+              </SubLabel>
+              <CodeBlock
+                code={`zingo-cli --data-dir /path/to/your/wallet/data`}
+              />
+            </div>
+            <div>
+              <SubLabel>
+                {qs?.startZingoCliCustom ?? "Start Zingo CLI (Custom)"}
+              </SubLabel>
+              <CodeBlock
+                code={`zingo-cli --chain mainnet --server http://127.0.0.1:8137 --data-dir /path/to/your/zaino/data`}
+              />
+            </div>
+            <div className="mt-[14px] p-[12px_14px] rounded-lg text-[13px] leading-relaxed bg-amber-500/[0.06] border border-amber-500/20 text-amber-600 dark:text-amber-400">
+              {qs?.zingoSyncWarning ??
+                "⚠ This will perform a full sync on first run, similar to lightwalletd."}
+            </div>
+          </SectionCard>
+        </div>
+      </section>
+
+      <ResourcesSection />
+    </div>
+  );
+};
+
+const ZakuraTab = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "40px" }}>
+      <section id="installing-zakura">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={1} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.installingZakura ?? "Installing Zakura"}
+          </h2>
+        </div>
+
+        <InfoBox>
+          <strong className="text-indigo-400 dark:text-indigo-300">
+            Zakura
+          </strong>{" "}
+          {qs?.zakuraInfo ??
+            "is a consensus-compatible Zcash full node written in Rust and built for high-performance syncing, pruning, snapshots, and zcashd-compatible RPC workflows."}
+        </InfoBox>
+
+        <div className="grid gap-5">
+          <SectionCard
+            title={qs?.recommendedInstaller ?? "Recommended Installer"}
+          >
+            <SubLabel>
+              {qs?.interactiveInstaller ?? "Interactive Installer"}
+            </SubLabel>
+            <CodeBlock
+              code={`curl -fsSL https://zakura.com/install.sh | bash`}
+            />
+            <div className="mt-[14px] p-[12px_14px] rounded-lg text-[13px] leading-relaxed bg-sky-500/[0.06] border border-sky-500/20 text-sky-700 dark:text-sky-400">
+              {qs?.zakuraInstallerNote ??
+                "The installer can set up either standard Zakura or its zcashd-compatible variant."}
+            </div>
+          </SectionCard>
+
+          <SectionCard title={qs?.cargoInstall ?? "Cargo Installation"}>
+            <SubLabel>
+              {qs?.dependencies ?? "Dependencies (Ubuntu/Debian)"}
+            </SubLabel>
+            <CodeBlock
+              code={`sudo apt update
+sudo apt install -y clang libclang-dev build-essential protobuf-compiler
+curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+source "$HOME/.cargo/env"`}
+            />
+
+            <SubLabel className="mt-[14px]">
+              {qs?.fromCratesIo ?? "From crates.io"}
+            </SubLabel>
+            <CodeBlock code={`cargo install --locked zakura`} />
+          </SectionCard>
+
+          <SectionCard title={qs?.fromSource ?? "From GitHub"}>
+            <SubLabel>{qs?.pinnedRelease ?? "Install Pinned Release"}</SubLabel>
+            <CodeBlock
+              code={`cargo install --git https://github.com/zakura-core/zakura --tag v1.2.0 zakura`}
+            />
+            <SubLabel className="mt-[14px]">
+              {qs?.verifyInstallation ?? "Verify Installation"}
+            </SubLabel>
+            <CodeBlock code={`zakurad --version`} />
+          </SectionCard>
+
+          <SectionCard title="Docker">
+            <SubLabel>{qs?.runLatestRelease ?? "Run Latest Release"}</SubLabel>
+            <CodeBlock
+              code={`docker run -d \
+  --name zakura \
+  -p 8233:8233 \
+  -v zakurad-cache:/home/zakura/.cache/zakura \
+  zakuracore/zakura:latest`}
+            />
+            <div className="mt-[14px] p-[12px_14px] rounded-lg text-[13px] leading-relaxed bg-amber-500/[0.06] border border-amber-500/20 text-amber-700 dark:text-amber-400">
+              {qs?.zakuraDockerNote ??
+                "Port 8233 exposes the Mainnet P2P service. Use port 18233 when configuring a Testnet container."}
+            </div>
+          </SectionCard>
+        </div>
+      </section>
+
+      <Divider />
+
+      <section id="running-zakura">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={2} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.runningZakura ?? "Running Zakura"}
+          </h2>
+        </div>
+
+        <div
+          className="grid gap-5"
+          style={{
+            gridTemplateColumns: "repeat(auto-fit, minmax(280px, 1fr))",
+          }}
+        >
+          <SectionCard title={qs?.commandLine ?? "Command Line"}>
+            <SubLabel>{qs?.startNode ?? "Start the Node"}</SubLabel>
+            <CodeBlock code={`zakurad start`} />
+
+            <SubLabel className="mt-[14px]">
+              {qs?.viewOptions ?? "View Available Options"}
+            </SubLabel>
+            <CodeBlock code={`zakurad start --help`} />
+          </SectionCard>
+
+          <SectionCard title={qs?.dockerManagement ?? "Docker Management"}>
+            <SubLabel>{qs?.viewLogs ?? "View Logs"}</SubLabel>
+            <CodeBlock code={`docker logs -f zakura`} />
+
+            <SubLabel className="mt-[14px]">
+              {qs?.stopAndRestart ?? "Stop and Restart"}
+            </SubLabel>
+            <CodeBlock
+              code={`docker stop zakura
+docker start zakura`}
+            />
+          </SectionCard>
+        </div>
+      </section>
+
+      <section className="rounded-[14px] p-8 mt-2 bg-sky-500/[0.03] dark:bg-sky-500/[0.05] border border-sky-500/15">
+        <h3 className="text-slate-900 dark:text-slate-100 text-[18px] font-bold mb-6">
+          {qs?.zakuraResources ?? "Zakura Resources"}
+        </h3>
+        <div
+          className="grid gap-2"
+          style={{
+            gridTemplateColumns: "repeat(auto-fit, minmax(240px, 1fr))",
+          }}
+        >
+          <ResourceLink href="https://zakura.com/">
+            {qs?.zakuraWebsite ?? "Zakura Website"}
+          </ResourceLink>
+          <ResourceLink href="https://github.com/zakura-core/zakura">
+            {qs?.sourceRepository ?? "Source Repository"}
+          </ResourceLink>
+          <ResourceLink href="https://github.com/zakura-core/zakura/releases">
+            {qs?.releases ?? "Releases"}
+          </ResourceLink>
+          <ResourceLink href="https://zakura.com/snapshots">
+            {qs?.snapshots ?? "Snapshots"}
+          </ResourceLink>
+        </div>
+      </section>
+    </div>
+  );
+};
+
+const ResourcesSection = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+  return (
+    <section className="rounded-[14px] p-8 mt-2 bg-sky-500/[0.03] dark:bg-sky-500/[0.05] border border-sky-500/15">
+      <h3 className="text-slate-900 dark:text-slate-100 text-[18px] font-bold mb-6">
+        {qs?.nextStepsResources ?? "Next Steps & Resources"}
+      </h3>
+      <div
+        className="grid gap-6"
+        style={{ gridTemplateColumns: "repeat(auto-fit, minmax(240px, 1fr))" }}
+      >
+        <div>
+          <p className="text-[11px] tracking-[0.12em] uppercase text-cyan-500 dark:text-cyan-400 mb-3 font-semibold">
+            {qs?.readyToBuild ?? "Ready to Build?"}
+          </p>
+          <div className="flex flex-col gap-2">
+            <ResourceLink href="https://zebra.zfnd.org/index.html">
+              {qs?.officialDocumentation ?? "Official Documentation"}
+            </ResourceLink>
+            <ResourceLink href="https://discord.gg/zcash">
+              {qs?.zcashDiscordCommunity ?? "Zcash Discord Community"}
+            </ResourceLink>
+            <ResourceLink href="https://github.com/zcash">
+              {qs?.githubRepositories ?? "GitHub Repositories"}
+            </ResourceLink>
+            <ResourceLink href="https://zips.z.cash/">
+              {qs?.zcashImprovementProposals ?? "Zcash Improvement Proposals"}
+            </ResourceLink>
+          </div>
+        </div>
+        <div>
+          <p className="text-[11px] tracking-[0.12em] uppercase text-cyan-500 dark:text-cyan-400 mb-3 font-semibold">
+            {qs?.needHelp ?? "Need Help?"}
+          </p>
+          <div className="flex flex-col gap-2">
+            <ResourceLink href="https://forum.zcashcommunity.com/">
+              {qs?.zcashCommunityForum ?? "Zcash Community Forum"}
+            </ResourceLink>
+            <ResourceLink href="https://github.com/zcash/zcash/issues">
+              {qs?.githubIssues ?? "GitHub Issues"}
+            </ResourceLink>
+            <ResourceLink href="https://zcash.readthedocs.io/en/latest/rtd_pages/troubleshooting.html">
+              {qs?.troubleshootingGuide ?? "Troubleshooting Guide"}
+            </ResourceLink>
+            <ResourceLink href="https://stackoverflow.com/questions/tagged/zcash">
+              {qs?.stackOverflow ?? "Stack Overflow"}
+            </ResourceLink>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+interface StackRow {
+  name: string;
+  href: string;
+  role: string;
+  language: string;
+  platforms: string;
+  status: "Active" | "Beta";
+  hardware: string;
+  rpc: string;
+}
+
+const STACK_ROWS: StackRow[] = [
+  {
+    name: "zebrad",
+    href: "https://github.com/ZcashFoundation/zebra",
+    role: "Full node",
+    language: "Rust",
+    platforms: "Linux · macOS · Windows (WSL)",
+    status: "Active",
+    hardware: "4 cores · 16 GB RAM · 300 GB disk",
+    rpc: "JSON-RPC (subset of zcashd methods)",
+  },
+  {
+    name: "Zakura",
+    href: "https://github.com/zakura-core/zakura",
+    role: "Full node (Zebra fork)",
+    language: "Rust",
+    platforms: "Linux · macOS · Windows (WSL)",
+    status: "Active",
+    hardware: "~11 GB disk pruned · snapshot bootstrap",
+    rpc: "JSON-RPC (zcashd compatibility mode)",
+  },
+  {
+    name: "Zaino",
+    href: "https://github.com/zingolabs/zaino",
+    role: "Indexer for light clients & explorers",
+    language: "Rust",
+    platforms: "Linux · macOS · Windows (WSL)",
+    status: "Beta",
+    hardware: "Light · runs beside a full node",
+    rpc: "gRPC (lightwalletd-compatible)",
+  },
+  {
+    name: "lightwalletd",
+    href: "https://github.com/zcash/lightwalletd",
+    role: "Light client server",
+    language: "Go",
+    platforms: "Linux · macOS · Windows (WSL)",
+    status: "Active",
+    hardware: "Light · needs a synced full node",
+    rpc: "gRPC",
+  },
+  {
+    name: "Zallet",
+    href: "https://github.com/zcash/wallet",
+    role: "Wallet backend (zcashd wallet replacement)",
+    language: "Rust",
+    platforms: "Linux · macOS · Windows (WSL)",
+    status: "Beta",
+    hardware: "Light · pairs with a Zebra node",
+    rpc: "JSON-RPC (zcashd wallet methods)",
+  },
+];
+
+const STATUS_STYLES: Record<StackRow["status"], string> = {
+  Active:
+    "bg-emerald-500/10 border-emerald-500/25 text-emerald-600 dark:text-emerald-400",
+  Beta: "bg-amber-500/10 border-amber-500/25 text-amber-600 dark:text-amber-400",
+};
+
+const StackComparison = () => {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+
+  const headers = [
+    qs?.compareColSoftware ?? "Software",
+    qs?.compareColRole ?? "Role",
+    qs?.compareColLanguage ?? "Language",
+    qs?.compareColPlatforms ?? "Platforms",
+    qs?.compareColStatus ?? "Status",
+    qs?.compareColHardware ?? "Hardware",
+    qs?.compareColRpc ?? "RPC / API",
+  ];
+
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "40px" }}>
+      <section id="stack-comparison">
+        <div className="flex items-center gap-[14px] mb-7">
+          <StepBadge number={1} />
+          <h2 className="text-slate-900 dark:text-slate-100 text-[22px] font-bold m-0">
+            {qs?.compareTitle ?? "Node & Client Comparison"}
+          </h2>
+        </div>
+
+        <SectionCard title={qs?.compareTableTitle ?? "At a Glance"}>
+          <p className="text-sm text-slate-600 dark:text-slate-400 leading-relaxed mt-0 mb-5">
+            {qs?.compareLead ??
+              "The Zcash stack is made up of a few separate tools. Use this table to work out which one your project needs, then follow the matching guide in the tabs above."}
+          </p>
+          <div className="overflow-x-auto">
+            <table className="w-full min-w-[840px] border-collapse text-[13px] text-left">
+              <thead>
+                <tr>
+                  {headers.map((h) => (
+                    <th
+                      key={h}
+                      className="py-[10px] px-3 text-[11px] tracking-[0.1em] uppercase font-semibold text-slate-500 dark:text-slate-500 border-b border-black/10 dark:border-white/10"
+                    >
+                      {h}
+                    </th>
+                  ))}
+                </tr>
+              </thead>
+              <tbody>
+                {STACK_ROWS.map((row) => (
+                  <tr
+                    key={row.name}
+                    className="border-b border-black/[0.06] dark:border-white/[0.06] last:border-b-0 align-top"
+                  >
+                    <td className="py-3 px-3 whitespace-nowrap">
+                      <a
+                        href={row.href}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                        className="text-sky-600 dark:text-sky-400 font-semibold no-underline hover:underline"
+                      >
+                        {row.name}
+                      </a>
+                    </td>
+                    <td className="py-3 px-3 text-slate-700 dark:text-slate-300">
+                      {row.role}
+                    </td>
+                    <td className="py-3 px-3 text-slate-500 dark:text-slate-400">
+                      {row.language}
+                    </td>
+                    <td className="py-3 px-3 text-slate-500 dark:text-slate-400">
+                      {row.platforms}
+                    </td>
+                    <td className="py-3 px-3 whitespace-nowrap">
+                      <span
+                        className={`inline-block px-2 py-[2px] rounded-full border text-[11px] font-semibold ${STATUS_STYLES[row.status]}`}
+                      >
+                        {row.status}
+                      </span>
+                    </td>
+                    <td className="py-3 px-3 text-slate-500 dark:text-slate-400">
+                      {row.hardware}
+                    </td>
+                    <td className="py-3 px-3 text-slate-500 dark:text-slate-400">
+                      {row.rpc}
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </div>
+          <p className="text-[12.5px] text-slate-500 dark:text-slate-500 leading-relaxed mt-4 mb-0">
+            <span className="text-amber-500 mr-1">⚠</span>
+            {qs?.compareFootnote ??
+              "zcashd reached end of life at block 3417100 on July 18, 2026 — new projects should build on the software above."}
+          </p>
+        </SectionCard>
+      </section>
+
+      <ResourcesSection />
+    </div>
+  );
+};
+
+// ── Main Page ────────────────────────────────────────────────────────────────
+
+const TABS = ["Zebrad", "Zakura", "Zaino", "Zingolib", "Compare"] as const;
+type Tab = (typeof TABS)[number];
+
+export default function QuickStartClient() {
+  const { t } = useLanguage();
+  const qs = t?.pages?.developersQuickStart;
+  const [activeTab, setActiveTab] = useState<Tab>("Zebrad");
+
+  const tabContent: Record<Tab, ReactNode> = {
+    Compare: <StackComparison />,
+    Zebrad: <ZebradTab />,
+    Zakura: <ZakuraTab />,
+    Zaino: <ZainodTab />,
+    Zingolib: <ZingolibTab />,
+  };
+
+  return (
+    <>
+      {/* Google Fonts */}
+      <style>{`
+        @import url('https://fonts.googleapis.com/css2?family=Space+Mono:wght@400;700&family=Syne:wght@400;500;600;700;800&display=swap');
+        * { box-sizing: border-box; }
+        ::-webkit-scrollbar { width: 6px; height: 6px; }
+        ::-webkit-scrollbar-track { background: #0a0e17; }
+        ::-webkit-scrollbar-thumb { background: #1e293b; border-radius: 3px; }
+      `}</style>
+
+      <div className="min-h-screen bg-white dark:bg-[#030712] text-slate-800 dark:text-slate-200">
+        {/* Hero */}
+        <div
+          className="relative overflow-hidden px-10 pt-20 pb-[72px] text-center
+            bg-gradient-to-b from-slate-50 to-white
+            dark:from-[#050d1a] dark:to-[#030712]"
+        >
+          {/* Grid overlay */}
+          <div
+            className="absolute inset-0 pointer-events-none"
+            style={{
+              backgroundImage: `
+                linear-gradient(rgba(14,165,233,0.06) 1px, transparent 1px),
+                linear-gradient(90deg, rgba(14,165,233,0.06) 1px, transparent 1px)
+              `,
+              backgroundSize: "40px 40px",
+              maskImage:
+                "radial-gradient(ellipse 80% 80% at 50% 50%, black 40%, transparent 100%)",
+            }}
+          />
+
+          {/* Glow */}
+          <div
+            className="absolute pointer-events-none"
+            style={{
+              top: "30%",
+              left: "50%",
+              transform: "translate(-50%, -50%)",
+              width: "600px",
+              height: "300px",
+              background:
+                "radial-gradient(ellipse, rgba(14,165,233,0.10) 0%, transparent 70%)",
+            }}
+          />
+
+          <div style={{ position: "relative" }}>
+            <div
+              className="inline-flex items-center gap-2 px-4 py-[6px] mb-7 rounded-full text-[12px] tracking-[0.12em] uppercase font-[500]
+                bg-sky-500/10 border border-sky-500/25 text-sky-600 dark:text-sky-400"
+              style={{ fontFamily: "'Space Mono', monospace" }}
+            >
+              <span
+                style={{
+                  width: 6,
+                  height: 6,
+                  borderRadius: "50%",
+                  background: "#22c55e",
+                  display: "inline-block",
+                  boxShadow: "0 0 6px #22c55e",
+                }}
+              />
+              {qs?.heroBadge ?? "Developer Quick Start"}
+            </div>
+
+            <h1
+              className="text-[clamp(36px,6vw,72px)] font-extrabold leading-[1.05] tracking-tight m-0 mb-5
+                bg-gradient-to-br from-slate-900 to-slate-600
+                dark:from-slate-100 dark:to-slate-400
+                bg-clip-text text-transparent"
+              style={{
+                WebkitBackgroundClip: "text",
+                WebkitTextFillColor: "transparent",
+              }}
+            >
+              {qs?.heroTitleLine1 ?? "Zcash Dev"}
+              <br />
+              {qs?.heroTitleLine2 ?? "Quick Start"}
+            </h1>
+
+            <p className="text-[17px] text-slate-600 dark:text-slate-500 max-w-[480px] mx-auto leading-[1.7]">
+              {qs?.heroSubtitle ??
+                "Get up and running with Zcash development. Choose your stack and follow the guide."}
+            </p>
+          </div>
+        </div>
+
+        {/* Tab bar */}
+        <div className="sticky top-0 z-10 bg-white/90 dark:bg-[#030712]/90 backdrop-blur-[12px] border-b border-black/[0.06] dark:border-white/[0.06] px-4 sam:px-6 lg:px-10">
+          <div className="max-w-[1024px] mx-auto flex gap-1 overflow-x-auto no-scrollbar">
+            {TABS.map((tab) => (
+              <button
+                key={tab}
+                onClick={() => setActiveTab(tab)}
+                className={`shrink-0 whitespace-nowrap px-4 py-3 lg:px-6 lg:py-4 bg-transparent border-0 border-b-2 cursor-pointer transition-all duration-200 text-[13px] lg:text-sm font-[inherit]
+          ${
+            activeTab === tab
+              ? "border-b-sky-500 text-slate-900 dark:text-slate-100 font-bold"
+              : "border-b-transparent text-slate-400 dark:text-slate-500 font-normal hover:text-slate-600 dark:hover:text-slate-300"
+          }`}
+                style={{ letterSpacing: "0.02em" }}
+              >
+                {tab}
+              </button>
+            ))}
+          </div>
+        </div>
+
+        {/* Content */}
+        <main className="max-w-[1024px] mx-auto px-4 imd:px-10 pt-12 pb-20">
+          {tabContent[activeTab]}
+        </main>
+      </div>
+    </>
+  );
+}
