+++ src/app/[locale]/explore/page.tsx
@@ -1,22 +1,37 @@
 import dynamic from "next/dynamic";
-import { genMetadata } from '@/lib/helpers';
-import { Metadata } from 'next';
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { Metadata } from "next";
+import { buildAlternates } from "@/lib/localeCoverage";
+import { routing } from "@/i18n/routing";
 
-export const metadata: Metadata = genMetadata({
-  title: "Explore Zcash",
-  url: "https://zechub.wiki/explore"
-})
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}): Promise<Metadata> {
+  const { locale } = await params;
+  const localePrefix =
+    locale && locale !== routing.defaultLocale ? `/${locale}` : "";
 
-const Explorer = dynamic(() => import("@/components/Explorer/Explorer"))
+  return genMetadata({
+    title: "Explore Zcash Ecosystem Directory | ZecHub",
+    description:
+      "Explore the interactive directory of Zcash projects, wallets, tools, community resources, and privacy technologies.",
+    url: `https://zechub.wiki${localePrefix}/explore`,
+    image: getBanner("start-here") || "/content-banners/bannerstart.png",
+    locale,
+    alternates: buildAlternates("/explore", locale, [locale]),
+  });
+}
 
+const Explorer = dynamic(() => import("@/components/Explorer/Explorer"));
 
 const Explore = async () => {
-
   return (
     <main className="">
       <Explorer />
     </main>
-  )
-}
+  );
+};
 
 export default Explore;
