+++ src/app/[locale]/hardware-wallets/page.tsx
@@ -1,12 +1,15 @@
-import { redirect } from 'next/navigation'
-import { genMetadata } from '@/lib/helpers';
-import { Metadata } from 'next';
+import { redirect } from "next/navigation";
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { Metadata } from "next";
 
 export const metadata: Metadata = genMetadata({
-    title: "Hardware walets",
-    url: "https://zechub.wiki/hardware-wallets"
-})
+  title: "Hardware Wallets for Zcash | ZecHub",
+  description:
+    "Secure your Zcash with hardware wallets supporting cold storage and shielded addresses.",
+  url: "https://zechub.wiki/hardware-wallets",
+  image: getBanner("using-zcash") || "/content-banners/usingzcash.png",
+});
 
 export default async function Page() {
-    redirect('/wallets')
+  redirect("/wallets");
 }
