+++ src/app/[locale]/mobile-wallets/page.tsx
@@ -1,12 +1,15 @@
-import { redirect } from 'next/navigation'
-import { genMetadata } from '@/lib/helpers';
-import { Metadata } from 'next';
+import { redirect } from "next/navigation";
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { Metadata } from "next";
 
 export const metadata: Metadata = genMetadata({
-    title: "Mobile walets",
-    url: "https://zechub.wiki/mobile-wallets"
-})
+  title: "Mobile Wallets for Zcash | ZecHub",
+  description:
+    "Find mobile wallets supporting Zcash (ZEC) shielded transactions on iOS and Android.",
+  url: "https://zechub.wiki/mobile-wallets",
+  image: getBanner("using-zcash") || "/content-banners/usingzcash.png",
+});
 
 export default async function Page() {
-    redirect('/wallets')
+  redirect("/wallets");
 }
