+++ src/app/[locale]/privacy/page.tsx
@@ -1,14 +1,27 @@
 import { Metadata } from "next";
-import { genMetadata } from "@/lib/helpers";
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { buildAlternates } from "@/lib/localeCoverage";
+import { routing } from "@/i18n/routing";
 
-export const metadata: Metadata = genMetadata({
-  title: "ZecHub — Privacy",
-  url: "https://zechub.wiki/privacy",
-});
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}): Promise<Metadata> {
+  const { locale } = await params;
+  const localePrefix =
+    locale && locale !== routing.defaultLocale ? `/${locale}` : "";
 
-// English-first. Fold into the translation-sync routine for the other locales.
-// This page states the *current* observable behaviour of the site; keep it in
-// sync when the observer surface changes (embeds, the AI assistant, hosting).
+  return genMetadata({
+    title: "Privacy Policy & Tracking Statement | ZecHub",
+    description:
+      "Learn how ZecHub protects your privacy. No trackers, no ad networks, no cookies, and zero surveillance on the open-source Zcash wiki.",
+    url: `https://zechub.wiki${localePrefix}/privacy`,
+    image: getBanner("privacy-tools") || "/content-banners/bannerprivacy.jpg",
+    locale,
+    alternates: buildAlternates("/privacy", locale, [locale]),
+  });
+}
 
 function Section({ title, children }: { title: string; children: React.ReactNode }) {
   return (
