+++ src/app/[locale]/visualizer/page.tsx
@@ -1,23 +1,45 @@
 import { VisualizerHub } from "@/components/visualizer/VisualizerHub";
-import { genMetadata } from "@/lib/helpers";
+import { genMetadata, getBanner } from "@/lib/helpers";
 import { Metadata } from "next";
-import { getDictionary } from '@/lib/getDictionary';
+import { getDictionary } from "@/lib/getDictionary";
+import { buildAlternates } from "@/lib/localeCoverage";
+import { routing } from "@/i18n/routing";
 
 type VisualizerDictionary = {
   pages?: {
     visualizer?: {
       title?: string;
+      description?: string;
     };
   };
 };
 
-export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }): Promise<Metadata> {
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}): Promise<Metadata> {
   const { locale } = await params;
-  const dict = (await getDictionary(locale)) as VisualizerDictionary;
+  const dict = (await getDictionary(locale).catch(() => ({}))) as VisualizerDictionary;
+  const localePrefix =
+    locale && locale !== routing.defaultLocale ? `/${locale}` : "";
+
+  const title = dict.pages?.visualizer?.title
+    ? `${dict.pages.visualizer.title} | ZecHub`
+    : "Interactive Zcash Visualizers & Tools | ZecHub";
+
+  const description =
+    dict.pages?.visualizer?.description ??
+    "Interactive cryptographic and blockchain visualizers for Zcash: zk-SNARKs, key derivation, consensus, hash functions, and shielded pools.";
+
   return genMetadata({
-    title: dict.pages?.visualizer?.title || "Zcash Visualizers",
-    url: "https://zechub.wiki/visualizer",
-  }) as Metadata;
+    title,
+    description,
+    url: `https://zechub.wiki${localePrefix}/visualizer`,
+    image: getBanner("zcash-tech") || "/content-banners/bannertech.jpg",
+    locale,
+    alternates: buildAlternates("/visualizer", locale, [locale]),
+  });
 }
 
 export default function VisualizerPage() {
