+++ src/app/[locale]/web-wallets/page.tsx
@@ -1,12 +1,15 @@
-import { redirect } from 'next/navigation'
-import { genMetadata } from '@/lib/helpers';
-import { Metadata } from 'next';
+import { redirect } from "next/navigation";
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { Metadata } from "next";
 
 export const metadata: Metadata = genMetadata({
-    title: "Web wallets",
-    url: "https://zechub.wiki/web-wallets"
-})
+  title: "Web Wallets for Zcash | ZecHub",
+  description:
+    "Discover web-based and browser extension wallets for interacting with Zcash.",
+  url: "https://zechub.wiki/web-wallets",
+  image: getBanner("using-zcash") || "/content-banners/usingzcash.png",
+});
 
 export default async function Page() {
-    redirect('/wallets')
+  redirect("/wallets");
 }
