+++ src/app/[locale]/zcash-projects/page.tsx
@@ -1,23 +1,46 @@
 import ZcashProjectsComp from "@/components/ZcashProjects/ZcashProjects";
 import { Metadata } from "next";
-import { genMetadata } from "@/lib/helpers";
-import { getDictionary } from '@/lib/getDictionary';
+import { genMetadata, getBanner } from "@/lib/helpers";
+import { getDictionary } from "@/lib/getDictionary";
+import { buildAlternates } from "@/lib/localeCoverage";
+import { routing } from "@/i18n/routing";
 
 type ProjectsDictionary = {
   pages?: {
     zcashProjects?: {
       title?: string;
+      description?: string;
     };
   };
 };
 
-export async function generateMetadata({ params }: { params: Promise<{ locale: string }> }): Promise<Metadata> {
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}): Promise<Metadata> {
   const { locale } = await params;
-  const dict = (await getDictionary(locale)) as ProjectsDictionary;
+  const dict = (await getDictionary(locale).catch(() => ({}))) as ProjectsDictionary;
+  const zp = dict.pages?.zcashProjects;
+  const localePrefix =
+    locale && locale !== routing.defaultLocale ? `/${locale}` : "";
+
+  const title = zp?.title
+    ? `${zp.title} | ZecHub`
+    : "Zcash Projects & Ecosystem Showcase | ZecHub";
+
+  const description =
+    zp?.description ??
+    "Discover active development projects, grants, privacy applications, and tools being built within the Zcash ecosystem.";
+
   return genMetadata({
-    title: dict.pages?.zcashProjects?.title || "Zcash Projects | ZecHub",
-    url: "https://zechub.wiki/zcash-projects",
-  }) as Metadata;
+    title,
+    description,
+    url: `https://zechub.wiki${localePrefix}/zcash-projects`,
+    image: getBanner("zcash-community") || "/content-banners/bannercommunity.jpg",
+    locale,
+    alternates: buildAlternates("/zcash-projects", locale, [locale]),
+  });
 }
 
 const ZcashProject = () => {
